face_login/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
click==7.1.2
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
dlib @ file:///D:/bld/dlib_1607165183225/work
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.0.4
imutils==0.5.3
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy @ file:///D:/bld/numpy_1604946253620/work
opencv-python==4.4.0.46
pandas @ file:///D:/bld/pandas_1609079363652/work
Pillow==8.0.1
psycopg2==2.8.6
pylint==2.6.0
python-dateutil==2.8.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1608904108784/work
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.12.1
